Extract Msf::DBManager::Import::FusionVM
MSP-11124
