+++ lib/msf/core/db_manager/import.rb
@@ -12,7 +12,6 @@ require 'uri'
 #
 
 require 'packetfu'
-require 'rex/parser/fusionvm_nokogiri'
 require 'rex/parser/mbsa_nokogiri'
 require 'rex/parser/nessus_xml'
 require 'rex/parser/netsparker_xml'
@@ -33,6 +32,7 @@ module Msf::DBManager::Import
   autoload :Burp, 'msf/core/db_manager/import/burp'
   autoload :CI, 'msf/core/db_manager/import/ci'
   autoload :Foundstone, 'msf/core/db_manager/import/foundstone'
+  autoload :FusionVM, 'msf/core/db_manager/import/fusion_vm'
   autoload :IP360, 'msf/core/db_manager/import/ip360'
   autoload :MsfXml, 'msf/core/db_manager/import/msf_xml'
   autoload :Qualys, 'msf/core/db_manager/import/qualys'
@@ -43,6 +43,7 @@ module Msf::DBManager::Import
   include Msf::DBManager::Import::Burp
   include Msf::DBManager::Import::CI
   include Msf::DBManager::Import::Foundstone
+  include Msf::DBManager::Import::FusionVM
   include Msf::DBManager::Import::IP360
   include Msf::DBManager::Import::MsfXml
   include Msf::DBManager::Import::Qualys
@@ -339,14 +340,6 @@ module Msf::DBManager::Import
     end
 
     raise DBImportError.new("Could not automatically determine file type")
-  end
-
-  def import_fusionvm_xml(args={})
-    args[:wspace] ||= workspace
-    bl = validate_ips(args[:blacklist]) ? args[:blacklist].split : []
-    doc = Rex::Parser::FusionVMDocument.new(args,self)
-    parser = ::Nokogiri::XML::SAX::Parser.new(doc)
-    parser.parse(args[:data])
   end
 
   def import_ip_list(args={}, &block)
